requirements.txt
Django
psycopg2-binary
python-binance
websocket-client
channels
dotenv
asgiref
plotly
pandas